plugin daapclient: split own code from upstream spydaap code
also sync spydaap with upstream code
